+++ priv/static/phoenix_live_view.esm.js
@@ -1435,6 +1435,11 @@ var DOMPatch = class {
           return el;
         },
         onNodeAdded: (el) => {
+          if (el instanceof HTMLImageElement && el.srcset) {
+            el.srcset = el.srcset;
+          } else if (el instanceof HTMLVideoElement && el.autoplay) {
+            el.play();
+          }
           if (dom_default.isNowTriggerFormExternal(el, phxTriggerExternal)) {
             externalFormTriggered = el;
           }